Update and rename README.md to run_away.py
This quick code is not very detailed only being made in a couple of hours but basic instructions include. The player is the red ball while the white balls are enemies. To start the game, chose between clicking h for hard or g for easy. For the hard mode, the age will reset if the player is stationary for too long, while easy does not do this. The player will only be able to move when the first two balls are sent out. To end the game, you will press space to display your high score and how long you survived in consol. The only way to win the game is by reaching a score of 2500 or about 45 seconds.